+++ airflow/dags/geoetl_airflow/utils/bigquery_utils.py
@@ -4,26 +4,28 @@ import logging
 from google.cloud import bigquery
 from google.api_core.exceptions import Conflict, NotFound, Forbidden
 
-from geoetl_airflow.file_utils import read_file
+from geoetl_airflow.utils.file_utils import read_file
 
 
 def create_dataset(client, dataset_name, project=None):
     dataset = client.dataset(dataset_name, project=project)
     try:
-        logging.info('Creating new dataset ...')
+        logging.info("Creating new dataset ...")
         dataset = client.create_dataset(dataset)
-        logging.info('New dataset created: ' + dataset_name)
+        logging.info("New dataset created: " + dataset_name)
     except Conflict as error:
-        logging.info('Dataset already exists')
+        logging.info("Dataset already exists")
     except Forbidden as error:
-        logging.info('User does not have bigquery.datasets.create permission in project')
+        logging.info(
+            "User does not have bigquery.datasets.create permission in project"
+        )
 
     return dataset
 
 
 def submit_bigquery_job(job, configuration):
     try:
-        logging.info('Creating a job: ' + json.dumps(configuration.to_api_repr()))
+        logging.info("Creating a job: " + json.dumps(configuration.to_api_repr()))
         result = job.result()
         logging.info(result)
         assert job.errors is None or len(job.errors) == 0
@@ -46,46 +48,48 @@ def read_bigquery_schema_from_json_recursive(json_schema):
     """
     result = []
     for field in json_schema:
-        if field.get('type').lower() == 'record' and field.get('fields'):
+        if field.get("type").lower() == "record" and field.get("fields"):
             schema = bigquery.SchemaField(
-                name=field.get('name'),
-                field_type=field.get('type', 'STRING'),
-                mode=field.get('mode', 'NULLABLE'),
-                description=field.get('description'),
-                fields=read_bigquery_schema_from_json_recursive(field.get('fields'))
+                name=field.get("name"),
+                field_type=field.get("type", "STRING"),
+                mode=field.get("mode", "NULLABLE"),
+                description=field.get("description"),
+                fields=read_bigquery_schema_from_json_recursive(field.get("fields")),
             )
         else:
             schema = bigquery.SchemaField(
-                name=field.get('name'),
-                field_type=field.get('type', 'STRING'),
-                mode=field.get('mode', 'NULLABLE'),
-                description=field.get('description')
+                name=field.get("name"),
+                field_type=field.get("type", "STRING"),
+                mode=field.get("mode", "NULLABLE"),
+                description=field.get("description"),
             )
         result.append(schema)
     return result
 
 
-def query(bigquery_client, sql, destination=None, priority=bigquery.QueryPriority.INTERACTIVE):
+def query(
+    bigquery_client, sql, destination=None, priority=bigquery.QueryPriority.INTERACTIVE
+):
     job_config = bigquery.QueryJobConfig()
     job_config.destination = destination
     job_config.priority = priority
-    logging.info('Executing query: ' + sql)
-    query_job = bigquery_client.query(sql, location='US', job_config=job_config)
+    logging.info("Executing query: " + sql)
+    query_job = bigquery_client.query(sql, location="US", job_config=job_config)
     submit_bigquery_job(query_job, job_config)
-    assert query_job.state == 'DONE'
+    assert query_job.state == "DONE"
 
 
 def create_view(bigquery_client, sql, table_ref):
     table = bigquery.Table(table_ref)
     table.view_query = sql
 
-    logging.info('Creating view: ' + json.dumps(table.to_api_repr()))
+    logging.info("Creating view: " + json.dumps(table.to_api_repr()))
 
     try:
         table = bigquery_client.create_table(table)
     except Conflict:
         # https://cloud.google.com/bigquery/docs/managing-views
-        table = bigquery_client.update_table(table, ['view_query'])
+        table = bigquery_client.update_table(table, ["view_query"])
     assert table.table_id == table_ref.table_id
     return table
